cleaned up recent tool checkin
